+++ pkg/cnab/bundleruntime/action_test.go
@@ -1,4 +1,4 @@
-package cnabprovider
+package bundleruntime
 
 import (
 	"encoding/json"
