+++ pkg/cnab/bundleruntime/credentials_test.go
@@ -1,4 +1,4 @@
-package cnabprovider
+package bundleruntime
 
 import (
 	"context"
