+++ pkg/cnab/bundleruntime/doc.go
@@ -1,0 +1,3 @@
+// Package bundleruntime implements the CNAB Bundle Runtime
+// and is used to run a single bundle.
+package bundleruntime
