+++ pkg/cnab/bundleruntime/main_test.go
@@ -1,4 +1,4 @@
-package cnabprovider
+package bundleruntime
 
 import (
 	"testing"
