+++ pkg/cnab/bundleruntime/testdata/bundle.json
