+++ pkg/cnab/bundleruntime/testdata/relocation-mapping.json
