--- pkg/cnab/provider/doc.go
@@ -1,1 +1,0 @@
-package cnabprovider
